chore: move plugins to the libs folder
