+++ src/libs/leaflet.rotatedmarker/leaflet.rotatedMarker.ts
